add options to convert multiple or single animal DLC datasets

+++ annolid/annotation/labelme2coco.py
@@ -235,7 +235,7 @@ def convert(input_annotated_dir,
                 points = np.asarray([x1, y1, x2, y1, x2, y2, x1, y2])
             elif shape_type == "circle":
                 (x1, y1), (x2, y2) = points
-                radius = ((x1-x2)**2 + (y1-y2)**2)**(1/2) + \
+                radius = int(((x1-x2)**2 + (y1-y2)**2)**(1/2)) +  \
                     np.random.choice(np.arange(0, 1, 0.1))
                 xs = x1 + (radius * np.cos(_angles))
                 ys = y1 + (radius * np.sin(_angles))
